Example of breaking InterMine API into multiple files.
